aftgangaglay: started working to fix VC Wall Wx

+++ src/agawwin.h
@@ -227,7 +227,8 @@ enum aga_result aga_mkwin(
 	win->height = height;
 
 	win->hwnd = CreateWindowA(AGA_CLASS_NAME, "Aft Gang Aglay", mask,
-							  CW_USEDEFAULT, CW_USEDEFAULT, width, height, 0, 0,
+							  CW_USEDEFAULT, CW_USEDEFAULT,
+							  (int) width, (int) height, 0, 0,
 							  env->module, 0);
 	if(!win->hwnd) return aga_winerr(__FILE__, "CreateWindowA");
 	if(!ShowWindow((void*) win->hwnd, SW_SHOWNORMAL)) {
